create lower level Block
